¯\_(ツ)_/¯

+++ src/models/SavingAccount.h
@@ -6,7 +6,8 @@
 class SavingAccount : public Account {
 private:
 	unsigned int limit_;
-	double interestRate_;
+	float interestRate_;
+
 public:
 	SavingAccount(const std::string&, const std::string&, const unsigned int&, const double&);
 	~SavingAccount();
